Add Dashboard view to break up container
