+++ frontend/src/containers/dashboard-container/index.spec.js
@@ -2,12 +2,12 @@ import { mount } from 'enzyme'
 import React from 'react'
 import { Provider } from 'react-redux'
 import configureMockStore from 'redux-mock-store'
-import * as actions from '../../actions'
-import GoogleAuthContext from '../../contexts/google'
-import SignInToken from '../../utils/signin-token'
+import * as actions from 'actions'
+import GoogleAuthContext from 'contexts/google'
+import SignInToken from 'utils/signin-token'
 import Dashboard from './index'
-jest.mock('../../utils/notify')
-jest.mock('../../utils/signin-token')
+jest.mock('utils/notify')
+jest.mock('utils/signin-token')
 
 describe('Dashboard', () => {
   beforeEach(() => {
