+++ frontend/src/views/dashboard/index.css
@@ -1,0 +1,3 @@
+.jukebox-dashboard {
+  padding: 20px;
+}
